refactor: moved/deleted/renamed files
created services object for retrieve/keep all services instances
