+++ src/commands/change-issue-assignee.ts
@@ -1,8 +1,7 @@
 import * as vscode from 'vscode';
 import { IssueItem } from '../explorer/item/issue-item';
-import { printErrorMessageInOutputAndShowAlert } from '../shared/log-utilities';
-import { selectAssignee } from '../shared/select-utilities';
-import state, { canExecuteJiraAPI, isWorkingIssue } from '../state/state';
+import state, { canExecuteJiraAPI, isWorkingIssue } from '../store/state';
+import services from '../services';
 
 export default async function changeIssueAssigneeCommand(issueItem: IssueItem): Promise<void> {
   try {
@@ -10,7 +9,7 @@ export default async function changeIssueAssigneeCommand(issueItem: IssueItem): 
       let issue = issueItem.issue;
       // verify if it's the current working issue
       if (!isWorkingIssue(issue.key)) {
-        let assignee = await selectAssignee(false, false, true, undefined);
+        let assignee = await services.selectValues.selectAssignee(false, false, true, undefined);
         if (!!assignee) {
           // call Jira API
           const res = await state.jira.setAssignIssue({ issueKey: issue.key, assignee: <string>assignee });
@@ -19,10 +18,10 @@ export default async function changeIssueAssigneeCommand(issueItem: IssueItem): 
       }
     } else {
       if (canExecuteJiraAPI()) {
-        printErrorMessageInOutputAndShowAlert('Use this command from Jira Plugin EXPLORER');
+        services.logger.printErrorMessageInOutputAndShowAlert('Use this command from Jira Plugin EXPLORER');
       }
     }
   } catch (err) {
-    printErrorMessageInOutputAndShowAlert(err);
+    services.logger.printErrorMessageInOutputAndShowAlert(err);
   }
 }
