+++ src/commands/change-issue-status.ts
@@ -1,8 +1,7 @@
 import * as vscode from 'vscode';
 import { IssueItem } from '../explorer/item/issue-item';
-import { printErrorMessageInOutputAndShowAlert } from '../shared/log-utilities';
-import { selectTransition } from '../shared/select-utilities';
-import state, { canExecuteJiraAPI, isWorkingIssue } from '../state/state';
+import state, { canExecuteJiraAPI, isWorkingIssue } from '../store/state';
+import services from '../services';
 
 export default async function changeIssueStatusCommand(issueItem: IssueItem): Promise<void> {
   try {
@@ -10,7 +9,7 @@ export default async function changeIssueStatusCommand(issueItem: IssueItem): Pr
       let issue = issueItem.issue;
       // verify if it's the current working issue
       if (!isWorkingIssue(issue.key)) {
-        const newTransitionId = await selectTransition(issue.key);
+        const newTransitionId = await services.selectValues.selectTransition(issue.key);
         if (newTransitionId) {
           // call Jira API
           const result = await state.jira.setTransition({
@@ -26,10 +25,10 @@ export default async function changeIssueStatusCommand(issueItem: IssueItem): Pr
       }
     } else {
       if (canExecuteJiraAPI()) {
-        printErrorMessageInOutputAndShowAlert('Use this command from Jira Plugin EXPLORER');
+        services.logger.printErrorMessageInOutputAndShowAlert('Use this command from Jira Plugin EXPLORER');
       }
     }
   } catch (err) {
-    printErrorMessageInOutputAndShowAlert(err);
+    services.logger.printErrorMessageInOutputAndShowAlert(err);
   }
 }
