+++ src/commands/favourites-filters.ts
@@ -1,14 +1,13 @@
 import { SEARCH_MODE } from '../shared/constants';
-import { printErrorMessageInOutputAndShowAlert } from '../shared/log-utilities';
-import { selectFavoriteFilters, selectIssue } from '../shared/select-utilities';
+import services from '../services';
 
 export default async function favouritesFiltersCommand(): Promise<void> {
   try {
-    const filter = await selectFavoriteFilters();
+    const filter = await services.selectValues.selectFavoriteFilters();
     if (filter) {
-      selectIssue(SEARCH_MODE.FAVOURITES_FILTERS, [filter.name, filter.jql]);
+      services.selectValues.selectIssue(SEARCH_MODE.FAVOURITES_FILTERS, [filter.name, filter.jql]);
     }
   } catch (err) {
-    printErrorMessageInOutputAndShowAlert(err);
+    services.logger.printErrorMessageInOutputAndShowAlert(err);
   }
 }
