+++ src/commands/issue-add-comment.ts
@@ -1,10 +1,8 @@
 import * as vscode from 'vscode';
 import { IssueItem } from '../explorer/item/issue-item';
-import { getConfigurationByKey } from '../shared/configuration';
 import { CONFIG } from '../shared/constants';
-import { printErrorMessageInOutputAndShowAlert } from '../shared/log-utilities';
-import { selectAssignee } from '../shared/select-utilities';
-import state, { canExecuteJiraAPI } from '../state/state';
+import state, { canExecuteJiraAPI } from '../store/state';
+import services from '../services';
 
 export default async function issueAddCommentCommand(issueItem: IssueItem): Promise<void> {
   try {
@@ -18,11 +16,11 @@ export default async function issueAddCommentCommand(issueItem: IssueItem): Prom
         // ask for assignee if there is one or more [@] in the comment
         const num = (text.match(new RegExp('[@]', 'g')) || []).length;
         for (let i = 0; i < num; i++) {
-          const assignee = await selectAssignee(false, false, true, undefined);
+          const assignee = await services.selectValues.selectAssignee(false, false, true, undefined);
           if (!!assignee) {
             text = text.replace('[@]', `[~${assignee}]`);
           } else {
-            printErrorMessageInOutputAndShowAlert('Abort command, wrong parameter.');
+            services.logger.printErrorMessageInOutputAndShowAlert('Abort command, wrong parameter.');
             return;
           }
         }
@@ -32,7 +30,7 @@ export default async function issueAddCommentCommand(issueItem: IssueItem): Prom
         // modal
         const action = await vscode.window.showInformationMessage('Comment created', 'Open in browser');
         if (action === 'Open in browser') {
-          const baseUrl = getConfigurationByKey(CONFIG.BASE_URL) || '';
+          const baseUrl = services.configuration.getConfigurationByKey(CONFIG.BASE_URL) || '';
           const url =
             `${baseUrl}/browse/${issue.key}` +
             `?focusedCommentId=${response.id}` +
@@ -43,10 +41,10 @@ export default async function issueAddCommentCommand(issueItem: IssueItem): Prom
       }
     } else {
       if (canExecuteJiraAPI()) {
-        printErrorMessageInOutputAndShowAlert('Use this command from Jira Plugin EXPLORER');
+        services.logger.printErrorMessageInOutputAndShowAlert('Use this command from Jira Plugin EXPLORER');
       }
     }
   } catch (err) {
-    printErrorMessageInOutputAndShowAlert(err);
+    services.logger.printErrorMessageInOutputAndShowAlert(err);
   }
 }
