+++ src/commands/open-issue.ts
@@ -1,9 +1,9 @@
 import * as vscode from 'vscode';
-import { getConfigurationByKey } from '../shared/configuration';
 import { CONFIG } from '../shared/constants';
+import services from '../services';
 
 export default async function openIssueCommand(issueId: string): Promise<void> {
   // open the issue in the browser
-  const url = `${getConfigurationByKey(CONFIG.BASE_URL)}/browse/${issueId}`;
+  const url = `${services.configuration.getConfigurationByKey(CONFIG.BASE_URL)}/browse/${issueId}`;
   vscode.commands.executeCommand('vscode.open', vscode.Uri.parse(url));
 }
