+++ src/explorer/item/divider-item.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import { DIVIDER } from '../../shared/constants';
-import { getIconsPath } from '../../shared/utilities';
+import services from '../../services';
 
 export class DividerItem extends vscode.TreeItem {
   constructor(label: string, collapsibleState?: vscode.TreeItemCollapsibleState) {
@@ -12,8 +12,8 @@ export class DividerItem extends vscode.TreeItem {
   }
 
   iconPath = {
-    light: getIconsPath(`light/${DIVIDER.file}`),
-    dark: getIconsPath(`dark/${DIVIDER.file}`)
+    light: services.utilities.getIconsPath(`light/${DIVIDER.file}`),
+    dark: services.utilities.getIconsPath(`dark/${DIVIDER.file}`)
   };
 
   contextValue = 'DividerItem';
