+++ src/explorer/item/limit-info.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import { STATUS_ICONS } from '../../shared/constants';
-import { getIconsPath } from '../../shared/utilities';
+import services from '../../services';
 
 export class LimitInfoItem extends vscode.TreeItem {
   constructor() {
@@ -24,8 +24,8 @@ export class LimitInfoItem extends vscode.TreeItem {
   }
 
   iconPath = {
-    light: getIconsPath(`light/${STATUS_ICONS.DEFAULT.file}`),
-    dark: getIconsPath(`dark/${STATUS_ICONS.DEFAULT.file}`)
+    light: services.utilities.getIconsPath(`light/${STATUS_ICONS.DEFAULT.file}`),
+    dark: services.utilities.getIconsPath(`dark/${STATUS_ICONS.DEFAULT.file}`)
   };
 
   contextValue = 'LimitInfoItem';
