+++ src/explorer/jira-explorer.ts
@@ -1,7 +1,6 @@
 import * as vscode from 'vscode';
-import { getConfigurationByKey } from '../shared/configuration';
-import { CONFIG, LOADING, LIST_MAX_RESULTS } from '../shared/constants';
-import state from '../state/state';
+import services from '../services';
+import { CONFIG, LIST_MAX_RESULTS, LOADING } from '../shared/constants';
 import { DividerItem } from './item/divider-item';
 import { FilterInfoItem } from './item/filter-info-item';
 import { IssueItem } from './item/issue-item';
@@ -9,6 +8,7 @@ import { LimitInfoItem } from './item/limit-info';
 import { LoadingItem } from './item/loading-item';
 import { NoResultItem } from './item/no-result-item';
 import { StatusItem } from './item/status-item';
+import state from '../store/state';
 
 export class JiraExplorer implements vscode.TreeDataProvider<IssueItem> {
   private _onDidChangeTreeData: vscode.EventEmitter<IssueItem | undefined> = new vscode.EventEmitter<IssueItem | undefined>();
@@ -25,7 +25,7 @@ export class JiraExplorer implements vscode.TreeDataProvider<IssueItem> {
   }
 
   async getChildren(element?: IssueItem): Promise<any[]> {
-    let project = await getConfigurationByKey(CONFIG.WORKING_PROJECT);
+    let project = await services.configuration.getConfigurationByKey(CONFIG.WORKING_PROJECT);
     const issues = state.issues;
     // generate all the item from issues saved in global state
     if (issues.length > 0) {
