+++ src/extension.ts
@@ -1,34 +1,24 @@
 import * as vscode from 'vscode';
 import commands from './commands';
-import { JiraExplorer } from './explorer/jira-explorer';
+import services from './services';
+import { GitIntegrationService } from './services/git-integration.service';
+import { StatusBarService } from './services/status-bar.service';
 import { CONFIG_NAME } from './shared/constants';
-import { GitIntegration } from './shared/git-integration';
-import { StatusBarManager } from './shared/status-bar';
-import state, { connectToJira } from './state/state';
+import state, { connectToJira } from './store/state';
 
 let channel: vscode.OutputChannel;
 
 export const activate = async (context: vscode.ExtensionContext): Promise<void> => {
   channel = vscode.window.createOutputChannel(CONFIG_NAME.toUpperCase());
-  context.subscriptions.push(channel);
-
-  const statusBar = new StatusBarManager();
-  const gitIntegration = new GitIntegration();
-  const jiraExplorer = new JiraExplorer();
-  vscode.window.registerTreeDataProvider('jiraExplorer', jiraExplorer);
-
-  // instances in the state
-  state.context = context;
   state.channel = channel;
-  state.statusBar = statusBar;
-  state.gitIntegration = gitIntegration;
-  state.jiraExplorer = jiraExplorer;
-
-  context.subscriptions.push(statusBar);
-  context.subscriptions.push(gitIntegration);
-
+  context.subscriptions.push(channel);
+  state.context = context;
+  services.gitIntegration = new GitIntegrationService();
+  services.statusBarManager = new StatusBarService();
+  vscode.window.registerTreeDataProvider('jiraExplorer', services.jiraExplorer);
+  context.subscriptions.push(services.statusBarManager);
+  context.subscriptions.push(services.gitIntegration);
   context.subscriptions.push(...commands.register());
-
   // create Jira Instance and try to connect
   await connectToJira();
 };
