+++ src/services/configuration.model.ts
@@ -1,6 +1,6 @@
 import { WorkspaceConfiguration } from 'vscode';
 
-export interface Configuration extends WorkspaceConfiguration {
+export interface IConfiguration extends WorkspaceConfiguration {
   baseUrl?: string;
   username?: string;
   workingProject?: string;
