+++ src/services/status-bar.service.ts
@@ -1,19 +1,18 @@
 import * as vscode from 'vscode';
 import { IWorkingIssue } from '../http/api.model';
-import state, { incrementStateWorkingIssueTimePerSecond } from '../state/state';
-import { getConfigurationByKey, getGlobalWorkingIssue, setGlobalWorkingIssue } from './configuration';
-import { CONFIG, NO_WORKING_ISSUE, TRACKING_TIME_MODE } from './constants';
-import { secondsToHHMMSS } from './utilities';
-const awayTimeout = parseInt(getConfigurationByKey(CONFIG.TRACKING_TIME_MODE_HYBRID_TIMEOUT) || '30', 10) * 60; // Default to 30 minutes
+import { CONFIG, NO_WORKING_ISSUE, TRACKING_TIME_MODE } from '../shared/constants';
+import state, { incrementStateWorkingIssueTimePerSecond } from '../store/state';
+import services from '.';
 
-export class StatusBarManager {
+export class StatusBarService {
   private workingProjectItem: vscode.StatusBarItem;
   private workingIssueItem: vscode.StatusBarItem;
   private intervalId: NodeJS.Timer | undefined;
-
+  private awayTimeout = 30 * 60; // Default to 30 minutes
   constructor() {
     this.workingIssueItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 100);
     this.workingProjectItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 200);
+    this.awayTimeout = parseInt(services.configuration.getConfigurationByKey(CONFIG.TRACKING_TIME_MODE_HYBRID_TIMEOUT) || '30', 10) * 60;
   }
 
   // setup working project item
@@ -22,13 +21,13 @@ export class StatusBarManager {
       return;
     }
     if (!project) {
-      project = (await getConfigurationByKey(CONFIG.WORKING_PROJECT)) || '';
+      project = (await services.configuration.getConfigurationByKey(CONFIG.WORKING_PROJECT)) || '';
     }
     this.workingProjectItem.tooltip = 'Set working project';
     this.workingProjectItem.command = 'jira-plugin.setWorkingProjectCommand';
     this.workingProjectItem.text = `$(clippy) ` + (!!project ? `Project: ${project}` : `Project: NONE`);
     this.workingProjectItem.show();
-    if (getConfigurationByKey(CONFIG.ENABLE_WORKING_ISSUE)) {
+    if (services.configuration.getConfigurationByKey(CONFIG.ENABLE_WORKING_ISSUE)) {
       this.updateWorkingIssueItem(true);
     }
   }
@@ -39,11 +38,11 @@ export class StatusBarManager {
 
   private workingIssueItemText(workingIssue: IWorkingIssue): string {
     return workingIssue.issue.key !== NO_WORKING_ISSUE.key
-      ? `Working Issue: ${workingIssue.issue.key || ''} $(watch) ${secondsToHHMMSS(workingIssue.trackingTime) || ''}` +
+      ? `Working Issue: ${workingIssue.issue.key || ''} $(watch) ${services.utilities.secondsToHHMMSS(workingIssue.trackingTime) || ''}` +
           (workingIssue.awayTime === 0
             ? ``
             : workingIssue.awayTime > 0
-            ? ` $(history) ${secondsToHHMMSS(awayTimeout - workingIssue.awayTime)}`
+            ? ` $(history) ${services.utilities.secondsToHHMMSS(this.awayTimeout - workingIssue.awayTime)}`
             : ` $(history) Away too long, issue timer paused`)
       : NO_WORKING_ISSUE.text;
   }
@@ -53,11 +52,11 @@ export class StatusBarManager {
     let issue;
     // verify stored working issue
     if (checkGlobalStore) {
-      issue = getGlobalWorkingIssue(state.context);
+      issue = services.configuration.getGlobalWorkingIssue(state.context);
       if (!!issue) {
         // if there is a stored working issue we will use it
         vscode.commands.executeCommand('jira-plugin.setWorkingIssueCommand', JSON.parse(issue), undefined);
-        setGlobalWorkingIssue(state.context, undefined);
+        services.configuration.setGlobalWorkingIssue(state.context, undefined);
         return;
       }
     }
@@ -67,7 +66,7 @@ export class StatusBarManager {
       this.startWorkingIssueInterval();
     } else {
       // if user select NO_WORKING_ISSUE clear the stored working issue
-      setGlobalWorkingIssue(state.context, undefined);
+      services.configuration.setGlobalWorkingIssue(state.context, undefined);
     }
     this.workingIssueItem.tooltip = this.workingIssueItemTooltip(state.workingIssue);
     this.workingIssueItem.command = 'jira-plugin.setWorkingIssueCommand';
@@ -85,8 +84,11 @@ export class StatusBarManager {
   public startWorkingIssueInterval(): void {
     this.clearWorkingIssueInterval();
     this.intervalId = setInterval(() => {
-      if (vscode.window.state.focused || getConfigurationByKey(CONFIG.TRACKING_TIME_MODE) === TRACKING_TIME_MODE.ALWAYS) {
-        if (getConfigurationByKey(CONFIG.TRACKING_TIME_MODE) === TRACKING_TIME_MODE.HYBRID) {
+      if (
+        vscode.window.state.focused ||
+        services.configuration.getConfigurationByKey(CONFIG.TRACKING_TIME_MODE) === TRACKING_TIME_MODE.ALWAYS
+      ) {
+        if (services.configuration.getConfigurationByKey(CONFIG.TRACKING_TIME_MODE) === TRACKING_TIME_MODE.HYBRID) {
           // If we are coming back from an away period catch up our logging time
           // If the away time was > awayTimeout, workingIssue.awayTime will be -1, so we won't log the away time.
           if (state.workingIssue.awayTime && state.workingIssue.awayTime > 0) {
@@ -97,10 +99,10 @@ export class StatusBarManager {
         }
         // Update as normal
         incrementStateWorkingIssueTimePerSecond();
-      } else if (getConfigurationByKey(CONFIG.TRACKING_TIME_MODE) === TRACKING_TIME_MODE.HYBRID) {
+      } else if (services.configuration.getConfigurationByKey(CONFIG.TRACKING_TIME_MODE) === TRACKING_TIME_MODE.HYBRID) {
         // If we are away from the Window capture the time, if it's less than our awayTimeout
         if (state.workingIssue.awayTime >= 0) {
-          if (awayTimeout - state.workingIssue.awayTime > 0) {
+          if (this.awayTimeout - state.workingIssue.awayTime > 0) {
             state.workingIssue.awayTime++;
           } else {
             // We've been away longer than the away timeout, we are probably working on something else
