+++ src/shared/document-link-provider.ts
@@ -1,13 +1,13 @@
 import * as vscode from 'vscode';
 import { IProject } from '../http/api.model';
-import { getConfigurationByKey } from './configuration';
 import { CONFIG } from './constants';
+import services from '../services';
 
 export class IssueLinkProvider implements vscode.DocumentLinkProvider {
   constructor(private projects: IProject[]) {}
 
   private get baseUrl(): string | undefined {
-    return getConfigurationByKey(CONFIG.BASE_URL);
+    return services.configuration.getConfigurationByKey(CONFIG.BASE_URL);
   }
 
   public provideDocumentLinks(document: vscode.TextDocument): vscode.ProviderResult<vscode.DocumentLink[]> {
